Calculate percentiles for a sliding window
Not exactly a full sliding window. I'm doing a window size of 200 and a
calculation step size of 10. This means that it'll start calculating
percentiles at 200 games for the window 0-200. Then again at 210 games
for the window 10-210. This helps to reduce the calculation overhead and
still provide enough data points on the graph to be interesting.